# backend-familleunie/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tontine Famille Unie API",
        "description": "API Laravel pour la gestion de la tontine Famille Unie",
        "contact": {
            "email": "admin@familleunie.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Statistiques globales du dashboard (ADMIN/TRESORIER)",
                "operationId": "2c29ca3389addbd18db1fff0f925cfd3",
                "responses": {
                    "200": {
                        "description": "Stats globales"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/contribution-types": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste des types de cotisations (ADMIN)",
                "operationId": "040bc3b94e4813b67bde2b858bde51d2",
                "responses": {
                    "200": {
                        "description": "Liste des types"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un type de cotisation (ADMIN)",
                "operationId": "bbef519e0d2ac063e5b2300dbee000e4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "amount"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "enum": [
                                            "weekly",
                                            "monthly"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Type créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/contribution-types/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un type de cotisation (ADMIN)",
                "operationId": "b23242a3553ba9f0d66df93460598947",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Modifier un type de cotisation (ADMIN)",
                "operationId": "62d367baf4ea46820be06d8d0018424a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type modifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/fund-types": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Liste des types de caisses (ADMIN)",
                "operationId": "20b7ef8fc0d0800a8d7692406f7e7677",
                "responses": {
                    "200": {
                        "description": "Liste des caisses"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer une caisse (ADMIN)",
                "operationId": "526e25074ba8ad0ab17b876b39965130",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "target_amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Caisse créée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/fund-types/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer une caisse (ADMIN)",
                "operationId": "2258ae9146a0c02396c988f5dacccdba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Caisse supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Modifier une caisse (ADMIN)",
                "operationId": "be3204022019b31a90aef282dc3cb940",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Caisse modifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/members": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un nouveau membre (ADMIN)",
                "operationId": "5495fce87318460263ec44db907a93c7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "roles"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Membre créé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/members/{id}/reset-password": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Réinitialiser le mot de passe d'un membre (ADMIN)",
                "operationId": "86eb109d442b58d1f0facb722baa384a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mot de passe réinitialisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/members/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Supprimer un membre (ADMIN)",
                "operationId": "91af855a1a2555d6d3f43f9045d880d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Membre supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Modifier les informations d'un membre (ADMIN)",
                "operationId": "07936bbdb32165895e5540aa7bfe586a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Membre mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/members/{id}/roles": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Modifier les rôles d'un membre (ADMIN)",
                "operationId": "bff1bf5893f0466bbb6112152f709201",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roles"
                                ],
                                "properties": {
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rôles mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre token Sanctum: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}